# openapi-sprint1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IQBF ULIMA API",
    "summary": "API transaccional para la gestión de IQBF",
    "description": "Contrato del Sprint 1: identidad, organización, insumos, presentaciones, densidades versionadas y auditoría.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Desarrollo local"
    }
  ],
  "paths": {
...
    "/api/catalogos/{catalogo}/{item_id}": {
      "patch": {
        "tags": [
          "Catálogos"
        ],
        "summary": "Editar o inactivar un valor de catálogo",
        "operationId": "update_catalog_item_api_catalogos__catalogo___item_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "catalogo",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "establecimientos",
                "carreras",
                "laboratorios",
                "ubicaciones",
                "investigadores",
                "roles"
              ],
              "type": "string",
              "title": "Catalogo"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "403": {
            "description": "Permiso denegado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Validación o regla de negocio",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CatalogItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "codigo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "estado": {
            "type": "string",
            "enum": [
              "ACTIVO",
              "INACTIVO"
            ],
            "title": "Estado"
          },
          "vigencia_desde": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vigencia Desde"
          },
          "vigencia_hasta": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vigencia Hasta"
          },
          "metadata": {
            "$ref": "#/components/schemas/CatalogMetadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "codigo",
          "nombre",
          "estado"
        ],
        "title": "CatalogItem"
      },
      "CatalogMetadata": {
        "properties": {
          "id_establecimiento": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Establecimiento"
          },
          "id_carrera": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Carrera"
          },
          "id_laboratorio": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Laboratorio"
          },
          "tipo": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PERSONA",
                  "AREA"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tipo"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          }
        },
        "type": "object",
        "title": "CatalogMetadata"
      },
      "CatalogUpdate": {
        "properties": {
          "codigo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9._-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo"
          },
          "nombre": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "estado": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ACTIVO",
                  "INACTIVO"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Estado"
          },
          "vigencia_hasta": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vigencia Hasta"
          },
          "id_establecimiento": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Establecimiento"
          },
          "id_carrera": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Carrera"
          },
          "id_laboratorio": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Laboratorio"
          },
          "tipo": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PERSONA",
                  "AREA"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tipo"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "title": "CatalogUpdate"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDACION",
              "ERROR_INTERNO",
              "ERROR_BASE_DATOS",
              "REGISTRO_DUPLICADO",
              "REFERENCIA_INVALIDA",
              "REGLA_NEGOCIO",
              "SALDO_INSUFICIENTE",
              "KARDEX_INMUTABLE",
              "SALDO_SOLO_VIA_KARDEX",
              "DENSIDAD_REQUERIDA",
              "VIGENCIA_SOLAPADA",
              "SESION_EXPIRADA",
              "TOKEN_INVALIDO",
              "NO_AUTENTICADO",
              "SESION_INVALIDA",
              "CREDENCIALES_INVALIDAS",
              "CUENTA_INACTIVA",
              "CUENTA_BLOQUEADA",
              "SIN_ROL",
              "PERMISO_DENEGADO",
              "CATALOGO_NO_EXISTE",
              "CATALOGO_SOLO_LECTURA",
              "REGISTRO_NO_EXISTE",
              "ESTABLECIMIENTO_REQUERIDO",
              "CODIGO_INSTITUCIONAL_REQUERIDO",
              "INSUMO_NO_EXISTE",
              "PRESENTACION_NO_EXISTE",
              "VIGENCIA_INVALIDA",
              "USUARIO_NO_EXISTE",
              "ROL_REQUERIDO",
              "ROL_INVALIDO",
              "ALCANCE_INVALIDO",
              "ALCANCE_GLOBAL_REQUERIDO",
              "MOTIVO_REQUERIDO",
              "DENSIDAD_LOTE_REQUERIDA",
              "INSUMO_INACTIVO"
            ],
            "title": "Code"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemFieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "request_id"
        ],
        "title": "Problem"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}